requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anthropic==0.23.1
anyio==4.3.0
asn1crypto==1.5.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloud-sql-python-connector==1.10.0
cryptography==42.0.5
dataclasses-json==0.6.4
deepdiff==7.0.1
dicttoxml==1.7.16
distro==1.9.0
docstring_parser==0.16
ecdsa==0.19.0
emoji==2.11.1
et-xmlfile==1.1.0
Faker==24.8.0
fastapi==0.110.1
fastapi-jwt==0.2.0
filelock==3.13.4
filetype==1.2.0
frozenlist==1.4.1
fsspec==2024.3.1
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-aiplatform==1.58.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpc-google-iam-v1==0.13.1
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6
joblib==1.4.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
langchain-core==0.2.11
langchain-google-cloud-sql-pg==0.6.1
langchain-google-vertexai==1.0.6
langchain-postgres==0.0.9
langchain-text-splitters==0.2.2
langdetect==1.0.9
langsmith==0.1.83
lxml==5.2.1
Mako==1.3.3
MarkupSafe==2.1.5
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
openpyxl==3.1.2
ordered-set==4.1.0
orjson==3.10.0
packaging==23.2
pandas==2.2.2
pdfminer.six==20231228
pg8000==1.31.1
pgvector==0.2.5
proto-plus==1.23.0
protobuf==4.25.3
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg-pool==3.2.1
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
pypdf==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.4.27
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.8.1
regex==2024.4.28
requests==2.31.0
rsa==4.9
scramp==1.4.5
shapely==2.0.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
sqlmodel==0.0.16
starlette==0.37.2
tabulate==0.9.0
tenacity==8.2.3
tokenizers==0.15.2
tqdm==4.66.2
types-protobuf==4.25.0.20240410
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
unstructured
unstructured-client
urllib3==2.2.1
uvicorn==0.29.0
wrapt==1.16.0
xmltodict==0.13.0
yarl==1.9.4
datasets==2.19.1
ragas==0.1.9